+++ nodemon.json
@@ -1,7 +1,8 @@
 {
   "watch": ["src"],
   "legacyWatch": true,
+  "delay": "500",
   "ext": "ts,html,scss,css",
   "ignore": ["src/public/stylesheets/*", ".test.ts"],
-  "exec": "npm run build:css && npm run build:js && node --experimental-modules `if [ $NODE_ENV = \"development\" ]; then echo \"--inspect=0.0.0.0 --nolazy\"; fi` ./dist/index.js"
+  "exec": "npm run build:css && npm run build:js && node --experimental-modules `if [ $NODE_ENV = \"development\" ]; then echo \"--inspect=0.0.0.0 --nolazy\"; fi` ./dist/server/index.js"
 }
