+++ package.json
@@ -6,17 +6,19 @@
   "scripts": {
     "dev": "npm run prisma:generate && nodemon",
     "start": "npm run prisma:generate && nodemon",
-    "start:prod": "nodemon dist/index.js",
+    "start:prod": "nodemon dist/server/index.js",
     "build": "npm run build:css && npm run prisma:generate && npm run build:js",
-    "build:js": "babel src --out-dir dist --extensions \".ts,.tsx\" --source-maps --copy-files",
-    "build:css": "sass src/server/styles.scss src/public/stylesheets/styles.css",
+    "build:js": "npm run build:js:server && npm run build:js:client",
+    "build:js:server": "BABEL_ENV=node babel src/server --out-dir dist/server --extensions \".ts,.tsx\" --copy-files",
+    "build:js:client": "BABEL_ENV=browser babel src/client --out-dir dist/client --extensions \".ts,.tsx\" --copy-files",
+    "build:css": "sass src/client/styles/styles.scss dist/client/styles/styles.css",
     "build:types": "tsc --emitDeclarationOnly",
     "type-check": "tsc --noEmit",
     "lint": "eslint .",
     "lint:fix": "eslint . --fix",
     "prettier-format": "prettier --config .prettierrc 'src/**/*.ts' --write",
-    "test": "jest --collect-coverage",
-    "test:watch": "jest --watch",
+    "test": "BABEL_ENV=node jest --collect-coverage",
+    "test:watch": "BABEL_ENV=node jest --watch",
     "prisma:generate": "npx prisma generate --schema ./src/server/models/schema.prisma",
     "prisma:migrate": "npx prisma migrate dev --schema ./src/server/models/schema.prisma",
     "prisma:reset": "npx prisma migrate reset --schema ./src/server/models/schema.prisma --force",
@@ -25,6 +27,7 @@
   "dependencies": {
     "@prisma/client": "^2.20.0",
     "@types/node-fetch": "^2.5.10",
+    "@types/url-search-params": "^1.1.0",
     "accessible-autocomplete": "^2.0.3",
     "aws-sdk": "^2.883.0",
     "body-parser": "^1.19.0",
@@ -41,6 +44,7 @@
     "pg": "^8.5.1",
     "prisma": "^2.20.0",
     "sqlstring": "^2.3.2",
+    "url-search-params": "^1.1.0",
     "winston": "^3.3.3"
   },
   "devDependencies": {
