+++ src/client/javascript/main.ts
@@ -1,0 +1,10 @@
+import URLSearchParams from "url-search-params";
+
+
+(() => {
+  // analytics
+  const {pathname} = window.location;
+  const urlParams = new URLSearchParams(window.location.search);
+
+  console.log({pathname, urlParams})
+})();
