+++ src/client/styles/styles.scss
@@ -1,4 +1,4 @@
-@import "../../node_modules/govuk-frontend/govuk/all";
+@import "../../../node_modules/govuk-frontend/govuk/all";
 
 .answers-box {
   background-color: govuk-colour("light-grey");
