+++ src/server/middlewares/__tests__/compression.test.ts
@@ -1,5 +1,5 @@
 import request from "supertest";
-import { server } from "../..";
+import { server } from "../../server";
 
 describe("Compress middleware", () => {
   test("is compressing when Accept-Encoding header is present", async () => {
