+++ src/server/middlewares/__tests__/helmet.test.ts
@@ -1,5 +1,5 @@
 import request from "supertest";
-import { server } from "../..";
+import { server } from "../../server";
 
 describe("Helmet middleware", () => {
   test("referrer-policy is correct", async () => {
