+++ src/server/middlewares/logger.ts
@@ -1,7 +1,7 @@
 import { RequestHandler } from "express";
 import expressWinston from "express-winston";
-import { logger } from "services/logger";
-import { DEBUG } from "config";
+import { logger } from "server/services/logger";
+import { DEBUG } from "server/config";
 
 export default (): RequestHandler => {
   return expressWinston.logger({
