+++ src/server/middlewares/static.ts
@@ -5,7 +5,7 @@ const ROOT = process.cwd();
 const NODE_MODULES = path.join(ROOT, "node_modules");
 const GOVUK_FRONTEND = path.join(NODE_MODULES, "govuk-frontend", "govuk");
 
-const publicFolderPath = path.join(__dirname, "..", "..", "public");
+const publicFolderPath = path.join(__dirname, "..", "..", "client");
 const govUKAssetsFolderPath = path.join(GOVUK_FRONTEND, "assets");
 const govUKAllJSPath = path.join(GOVUK_FRONTEND, "all.js");
 const accessibleAutoCompletePath = path.join(
