+++ src/server/models/helpers.ts
@@ -1,5 +1,5 @@
 import { upperFirst, isNumber } from "lodash";
-import { logger } from "services/logger";
+import { logger } from "server/services/logger";
 import { db } from "./database";
 
 const countriesWithData = ["Thailand", "France", "Italy", "Spain"];
