+++ src/server/models/types.d.ts
@@ -1,5 +1,5 @@
 import Prisma from "@prisma/client";
-import { countriesList } from "services/metadata"
+import { countriesList } from "server/services/metadata"
 
 export type CountriesWithData = Extract<CountryName, "Thailand" | "France" | "Italy" | "Spain">;
 export type CountryName = typeof countriesList[number]["text"];
