+++ src/server/services/__mocks__/aws-sdk.ts
