+++ src/server/services/location.ts
@@ -3,7 +3,7 @@ import {
   LOCATION_SERVICE_ACCESS_KEY,
   LOCATION_SERVICE_SECRET_KEY,
   LOCATION_SERVICE_INDEX_NAME,
-} from "config";
+} from "server/config";
 import { logger } from "./logger";
 
 const INDEX_PARAMS = {
