+++ src/server/views/layout.html
@@ -8,6 +8,7 @@
 {% block head %}
 
   {% if GA_TRACKING_ID and (not isLocalHOST) %}
+  {# {% if GA_TRACKING_ID %} #}
     <script async src="https://www.googletagmanager.com/gtag/js?id={{ GA_TRACKING_ID }}"></script>
     <script>
       window.dataLayer = window.dataLayer || [];
@@ -21,7 +22,7 @@
   <link rel="preload" as="font" href="/assets/fonts/bold-b542beb274-v2.woff2" type="font/woff2" crossorigin="anonymous">
   <link rel="preload" as="font" href="/assets/fonts/light-94a07e06a1-v2.woff2" type="font/woff" crossorigin="anonymous">
   <link rel="preload" as="font" href="/assets/fonts/light-f591b13f7d-v2.woff" type="font/woff" crossorigin="anonymous">
-  <link href="/assets/stylesheets/styles.css" rel="stylesheet" />
+  <link href="/assets/styles/styles.css" rel="stylesheet" />
   <link href="/assets/accessible-autocomplete.min.css" rel="stylesheet" />
   <meta name="Description" content="Find {{ serviceLabel if serviceLabel else "a Professional Service"}} {{ "in " + _.startCase(country) if country else "abroad"}} - GOV.UK">
   
@@ -54,6 +55,7 @@
   <script>
     document.body.className = ((document.body.className) ? document.body.className + ' js-enabled' : 'js-enabled');
   </script>
+  <script src="/assets/javascript/main.js"></script>
 {% endblock %}
